requirements.txt
Django==2.1.3
django-thumbnail-maker==0.0.6
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
sorl-thumbnail==12.5.0
